backend/requirements.txt
fastapi
sqlalchemy
pydantic
pydantic[email]
uvicorn
pytest
httpx
psycopg2
dotenv
alembic
python-slugify